Don't print messages to stdout when importing Pygame
"Why are you doing this, Carter? Don't you want the Pygame project to succeed?"
Well, yes, but I'll be distributing this build of Pygame to children (readers
of Hello World! Third Ed). Directing kids to visit a chatroom we don't control
is a bad idea that could lead to legal issues.

Sending kids to the Pygame Discord channel would probably also be bad for the
Pygame project. A large influx of first-time programmers would be an "Eternal
September"-style disruption to the more experienced developers and project
maintainers in the Pygame community.